DocQA-MS/rag-engine/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pika==1.3.2
redis==5.0.1
psycopg2-binary==2.9.9
faiss-cpu==1.7.4
numpy==1.24.3
sentence-transformers>=2.3.0
huggingface-hub>=0.19.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
pydantic==2.5.0
rank-bm25==0.2.2
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
transformers==4.35.2
torch==2.1.0
